# Fix priority inversion in main thread tile loading