… Cesium3DTilesSelection/test/TestTilesetSelectionAlgorithm.cpp