Fixes #602 
This is a PR into #599 so merge that first.

Measuring load times for the Melbourne Photogrammetry tileset, in the default view in 02_CesiumMelbourne in Cesium for Unity Samples. Looking at the time from when the tileset is created until the last tile that will be selected _is_ selected. In other words, when the `Rendered` count stops changing. I pressed "Refresh Tileset" 5 times, waiting for it to finish refining in between each, then did it 10 more times and recorded the time for each. So the cache was warmed up in all cases.

Before this PR:

| *Refresh Number* | *Time (ms)* |
|---------------------|----------|
| 1 | 5602 |
| 2 | 6166 |
| 3 | 7007 |
| 4 | 6938 |
| 5 | 6217 |
| 6 | 8885 |
| 7 | 5161 |
| 8 | 6649 |
| 9 | 6769 |
| 10 | 6577 |
| Min | 5161 |
| Max | 8885 |
| Median | 6613 |
| Median reloads per hour | 544 |

After this PR:

| *Refresh Number* | *Time (ms)* |
|---------------------|----------|
| 1 | 1869 |
| 2 | 1742 |
| 3 | 2999 |
| 4 | 3055 |
| 5 | 2139 |
| 6 | 2091 |
| 7 | 2207 |
| 8 | 3224 |
| 9 | 2302 |
| 10 | 1911 |
| Min | 1742 |
| Max | 3224 |
| Median | 2173  |
| Median reloads per hour | 1657 |

So based on the median times, this branch is over 200% faster!

Note that the times are kind of all over the place in both branches. I don't know why there is so much variability, and I'm a bit concerned about it. Still, the trend is obvious!