@kring 2023-03-10
Here is a similar timing for Cesium World Terrain in 01_CesiumWorld.

Before this PR:

| *Refresh Number* | *Time (ms)* |
|---------------------|--------------|
| 1 | 985 |
| 2 | 966 |
| 3 | 952 |
| 4 | 1090 |
| 5 | 1002 |
| 6 | 1103 |
| 7 | 920 |
| 8 | 1035 |
| 9 | 936 |
| 10 | 953 |
| Min | 920 |
| Max | 1103 |
| Median | 975.5 |
| Median reloads per hour |  3690 |

After this PR:

| *Refresh Number* | *Time (ms)* |
|---------------------|--------------|
| 1 | 849 |
| 2 | 998 |
| 3 | 810 |
| 4 | 882 |
| 5 | 869 |
| 6 | 919 |
| 7 | 801 |
| 8 | 855 |
| 9 | 822 |
| 10 | 948 |
| Min | 801 |
| Max | 998 |
| Median | 862 |
| Median reloads per hour | 4176 |

A more modest improvement here, about 13% faster.