@kring 2023-03-16
The trouble with inserting them in sorted order in the first place is that it makes the "kick" mechanism much harder (or at least slower) to implement. When all the tiles for a subtree are at the end of the load queue, it's easy and cheap to remove them all.