@javagl 2023-03-17
I'm waaay out of the code (and wouldn't even claim to have been "in" the code at any point in time), and only vaguely remember the "kick" mechanism. But even if the tile load tasks have to be maintained in a linear list in order to make the `kick` easy and efficient, one could consider options to _avoid_ the `sort` call. 

Specifically: When there's a strong invariant that the queue is always sorted, then a 'sorted insert' of new elements could be implemented trivially, and done in O(n) time. (Of course, that "sorted insert" would not be implemented on a `vector`, but a `linked list` - an indexed insert for a vector has O(n) on its own, whereas it's O(1) for a linked list).

Whether or not that brings any benefit is hard to say, and depends on all the obvious factors: How many elements are in the queue, how many are added, how many are removed (individually or 'kicked' in bulk). But it might be worth considering, at least.