@kring 2023-03-18
To make the kick easy, the list has to be maintained in _traversal_ order (i.e. depth-first, pre-order). But for loading, they need to be in _priority_ order (which is very different from traversal order). I get what you're saying about inserting in sorted order, and that being more efficient (possible cache-thrashing of linked list data structures aside). I just haven't quite seen a way to do it efficiently yet given the need for the two different orderings for different purposes.