@javagl 2023-03-18
OK, beyond that, I could only brainstorm about this on an even higher level, at least as long as I don't refresh my memories. (That _would_ start with looking at the "kick" thing, and how that can operate on a list in traversal order, when the actual list is _sorted_ occasionally, thus "destroying" any other order that it might have had). This is not going to happen soon, but maybe I can at least allocate a chunk of time to update https://github.com/javagl/cesium-cpp to work with the latest state, and do some "smaller" benchmarks at some point...