@joseph-kaile 2023-03-22
I created a test to time Tileset Refresh and here are the results:

(Melbourne -- No Priority Inversion, Forbid holes, preload ancestors, preload siblings)

4.976 

4.824

4.756

4.787


4.817

-------------------------------------------------------

(Melbourne -- Forbid holes, preload ancestors, preload siblings)

5.034

5.093

4.939

5.093

4.982

-------------------------------------------------------

(Melbourne -- No Priority Inversion, No Forbid holes, No preload ancestors, No preload siblings)


2.979

2.995

3.089


-------------------------------------------------------

(Melbourne -- No Forbid holes, No preload ancestors, No preload siblings)

3.015

2.970

3.194

-------------------------------------------------------

This is the scene that it loads:

![HighresScreenshot00000](https://user-images.githubusercontent.com/97980867/227021039-7717480a-b072-4c0a-9d61-f0468a9a8139.png)


And the script to time the Tileset refresh:

```
// Fill out your copyright notice in the Description page of Project Settings.

#pragma once

#include "CoreMinimal.h"
#include "GameFramework/Actor.h"
#include "Cesium3DTileset.h"
#include "MyActor.generated.h"

UCLASS()
class MYPROJECT_API AMyActor : public AActor
{
	GENERATED_BODY()
	
public:	
	// Sets default values for this actor's properties
	AMyActor();

	UPROPERTY(EditAnywhere)
		ACesium3DTileset* Tileset;

protected:
	// Called when the game starts or when spawned
	virtual void BeginPlay() override;

	UFUNCTION()
	void ProcessLoad();

	int NumRefreshes = 0;

	float AccumulatedTime = 0;

	float NextLoadCountDown = 10.0f;

	bool FirstLoad = true;

	const int32 TotalNumRefreshes = 10;

	FTimerHandle NextLoadHandle;

	float CurrentTimeElapsed = 0;

	UFUNCTION()
		void OnCountDown();

};


// Fill out your copyright notice in the Description page of Project Settings.


#include "MyActor.h"

// Sets default values
AMyActor::AMyActor()
{
 	// Set this actor to call Tick() every frame.  You can turn this off to improve performance if you don't need it.
	PrimaryActorTick.bCanEverTick = true;

}

// Called when the game starts or when spawned
void AMyActor::BeginPlay()
{
	Super::BeginPlay();
	Tileset->MaximumCachedBytes = 0;
	FScriptDelegate delegate;
	delegate.BindUFunction(this, FName("ProcessLoad"));
	Tileset->OnTilesetLoaded.Add(delegate);

	OnCountDown();
}
void AMyActor::ProcessLoad()
{
	CurrentTimeElapsed = GetWorldTimerManager().GetTimerElapsed(NextLoadHandle);

	UE_LOG(LogTemp, Warning, TEXT("Time elapsed: %f"), CurrentTimeElapsed);
}

void AMyActor::OnCountDown()
{

	if (FirstLoad) {
		FirstLoad = false;
		GetWorld()->GetTimerManager().ClearTimer(NextLoadHandle);
		GetWorld()->GetTimerManager().SetTimer(NextLoadHandle, this, &AMyActor::OnCountDown, 10.0f, true);
		Tileset->RefreshTileset();
	}
	else {
		AccumulatedTime += CurrentTimeElapsed;
		CurrentTimeElapsed = 0;

		if (NumRefreshes++ == TotalNumRefreshes) {

			float AverageTime = AccumulatedTime / float(TotalNumRefreshes);
			UE_LOG(LogTemp, Warning, TEXT("Average Time: %f"), AverageTime);
			Tileset->Destroy();
		}
		else {
			GetWorld()->GetTimerManager().ClearTimer(NextLoadHandle);
			GetWorld()->GetTimerManager().SetTimer(NextLoadHandle, this, &AMyActor::OnCountDown, NextLoadCountDown, true);
			Tileset->RefreshTileset();
		}
	}
}
```